requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
httpx
pytest
python-jose[cryptography]
alembic
psycopg[binary]
